move navigation and sidebar to its own module
