+++ AMW_angular/io/src/app/navigation/navigation.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { NavigationComponent } from './navigation.component';
+import { NavigationSidebarComponent } from './navigation-sidebar.component';
+
+@NgModule({
+  declarations: [NavigationComponent, NavigationSidebarComponent],
+  exports: [NavigationComponent, NavigationSidebarComponent],
+  imports: [CommonModule]
+})
+export class NavigationModule {}
